requirements.txt
requests>=2.28.0
pyyaml>=6.0
click>=8.0.0
jinja2>=3.1.0
pandas>=1.5.0
openpyxl>=3.0.0
python-dateutil>=2.8.0
colorama>=0.4.0
tabulate>=0.9.0
cryptography>=3.4.0
urllib3>=1.26.0
certifi>=2022.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
